fix: split file of disable-artsan-controller for multple controlers with dto for UUID
